requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai>=1.37.0
faiss-cpu
tiktoken
numpy
